requirements.txt
bidict==0.22.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
et-xmlfile==1.1.0
eventlet==0.33.3
Flask==2.2.2
Flask-SocketIO==5.3.2
greenlet==2.0.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
netifaces==0.10.6
openpyxl==3.1.1
PyJWT==2.6.0
pymongo==4.3.3
python-engineio==4.3.4
python-socketio==5.7.2
six==1.16.0
socketio==0.2.1
Werkzeug==2.2.3
zipp==3.13.0
